adding docs back into project directory
