+++ opencb/models/__init__.py
@@ -4,4 +4,4 @@ from .d16a1gs import d16a1gs
 from .horse import horse
 from .grizzlybear import grizzlybear
 from .general import general
-from .general_dev import general_dev
+from .general import general_dev
