+++ requirements.txt
@@ -1,0 +1,2 @@
+sphinx==7.3.7
+sphinx-autoapi==3.0.0
