+++ source/index.rst
@@ -1,0 +1,20 @@
+.. OpenCB documentation master file, created by
+   sphinx-quickstart on Fri May 31 20:48:38 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to OpenCB's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
